ARM: imx6: revert changes done by v5.4.52 merge
Commit 567092c5f634cbd7b29472b1063d442b62b3b49b (Merge tag 'v5.4.52'
into 5.4-1.0.0-imx) ended up bringing the pm-imx6.c content from
upstream, as a consequence of applying
10533390da7cf1e2750c63dae44cacfa5964dce1 (ARM: imx6: add missing
put_device() call in imx6q_suspend_init()), making the current version
incompatible with the rest of the BSP-based tree.

This was noticed when building the kernel for a iMX6-based device that
used pm-imx6, as the build failed due missing imx6_pm_map_io.

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>