kbuild: filter only valid dtb files to construct complex dtb
The include files from dts are included as real-prereq files to
arguments when a complex dtb files is built. fdtoverlay command is
used for such building and *.h input parameters break the command
working.

Introduce a filter for dtb/dtbo files only to build a correct input
parameters list for complex dtb.

Signed-off-by: Oleksandr Suvorov <oleksandr.suvorov@foundries.io>