drivers: gpu: drm: msm: registers: improve reproducibility
Backport commit 874746eaa341 ("drivers: gpu: drm: msm: registers:
improve reproducibility") from linux-yocto repository.

Fix-up the generated name in the inner loop, previous patch version used
the same last name from the first loop in the second loop that produced
the actual output.

Original commit description:
----
The files generated by gen_header.py capture the source path to the
input files and the date.  While that can be informative, it varies
based on where and when the kernel was built as the full path is
captured.

Since all of the files that this tool is run on is under the drivers
directory, this modifies the application to strip all of the path before
drivers.  Additionally it prints <stripped> instead of the date.

Both changes solve the reproducibility issue.

Upstream-Status: Inappropriate
----

Signed-off-by: Ryan Eatmon <reatmon@ti.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>
Signed-off-by: Andrey Zhizhikin <andrey.z@gmail.com>